clean up Services.js
