+++ src/components/services/Services.js
@@ -3,17 +3,9 @@ import { services, aboutMe } from '../../text/services';
 import styles from './Services.css';
 import Title from '../title/Title';
 import ContentDevider from '../content-devider/ContentDevider';
+import { makeServicesList } from './makeServicesList';
 
 export default function Services() {
-  const servicesList = services.map((content, i) => (
-    <li className={styles.li} key={i}>
-      <>
-        <h2>{content.title}</h2>
-        <p className={styles.devider}>~</p>
-        <p>{content.content}</p>
-      </>
-    </li>
-  ));
   return (
     <>
       <main
@@ -21,7 +13,7 @@ export default function Services() {
         className={[styles.servicesPage, 'servicesPage'].join(' ')}
       >
         <Title title="services" />
-        <ul className={styles.servicesList}>{servicesList}</ul>
+        <ul className={styles.servicesList}>{makeServicesList(services)}</ul>
         <p className={styles.aboutMe}>{aboutMe}</p>
       </main>
       <ContentDevider />
